Update main.py
Now able to add just the year to MP4 metadata, but it will likely be necessary to revisit this idea as a whole, as it appears there is some confusion with how creation dates are tagged onto video files.

Ultimately, I want this metadata to feed into a photo/video manager like Immich, so we just need to add metadata that will allow the application to know how to sort the file. (https://github.com/immich-app/immich/discussions/1947)